# swagger.yaml
openapi: 3.0.0
info:
  title: BiteBuddies API
  description: API documentation for BiteBuddies application services
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Account Service
  - url: http://localhost:5001
    description: Availability Service
  - url: http://localhost:5002
    description: Restaurant Service
  - url: http://localhost:5006
    description: Transcription Service
  - url: http://localhost:5007
    description: Chatbot Service
  - url: http://localhost:5010
    description: Search Service
...
paths:
...
  /restaurants:
    get:
      tags:
        - Restaurant
      summary: Get restaurants by region
      description: Retrieve a list of restaurants in a specific region
      parameters:
        - name: region
          in: query
          required: true
          description: Region name
          schema:
            type: string
      responses:
        "200":
          description: List of restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "400":
          description: Bad request - missing region parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    post:
      tags:
        - Restaurant
      summary: Add a new restaurant
      description: Create a new restaurant entry in the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - address
                - region
                - latitude
                - longitude
              properties:
                name:
                  type: string
                  description: Restaurant name
                address:
                  type: string
                  description: Restaurant address
                region:
                  type: string
                  description: Region where restaurant is located
                latitude:
                  type: number
                  description: Latitude coordinate
                longitude:
                  type: number
                  description: Longitude coordinate
                cuisine:
                  type: string
                  description: Type of cuisine
                price_range:
                  type: string
                  description: Price range (low, medium, high)
      responses:
        "201":
          description: Restaurant added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
        "400":
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /restaurant/id:
    get:
      tags:
        - Restaurant
      summary: Get restaurant ID by name
      description: Retrieve a restaurant's ID using its name
      parameters:
        - name: name
          in: query
          required: true
          description: Restaurant name
          schema:
            type: string
      responses:
        "200":
          description: Restaurant ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        "400":
          description: Bad request - missing name parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "404":
          description: Restaurant not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
  /restaurants/get_by_name:
    post:
      tags:
        - Restaurant
      summary: Get restaurant by name
      description: Retrieve restaurant details using its name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Restaurant name
      responses:
        "200":
          description: Restaurant details
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
        "400":
          description: Bad request - missing restaurant name
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "404":
          description: Restaurant not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /restaurants/nearby:
    post:
      tags:
        - Restaurant
      summary: Get nearby restaurants
      description: Find restaurants within a specified radius from a location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - latitude
                - longitude
              properties:
                latitude:
                  type: number
                  description: Latitude coordinate
                longitude:
                  type: number
                  description: Longitude coordinate
                radius_km:
                  type: number
                  description: Search radius in kilometers
                  default: 2.0
      responses:
        "200":
          description: List of nearby restaurants
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      type: object
                  count:
                    type: integer
                  radius_km:
                    type: number
        "400":
          description: Bad request - missing latitude or longitude
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  error:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  error:
                    type: string
  /restaurants/sample_data:
    get:
      tags:
        - Restaurant
      summary: Add sample restaurant data
      description: Add sample restaurant data to the database for testing
      responses:
        "200":
          description: Sample data status
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                  restaurant_ids:
                    type: array
                    items:
                      type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  error:
                    type: string

  # =========== TRANSCRIPTION SERVICE ===========
...